Split out gm_16A2
lbl_8046B488 appears to be the start of a file's bss section, based on
it being used as a base offset for bss accesses. Since there is already
float duplication in 1602 that will require splitting, let's split out a
new file based on this bss boundary.
