+++ src/melee/if/ifstock.c
@@ -10,8 +10,7 @@
 
 #include "gm/forward.h"
 
-#include "gm/gm_1601.h"
-#include "gm/gm_16AE.h"
+#include "gm/gm_unsplit.h"
 #include "gm/gmregclear.h"
 #include "gm/types.h"
 #include "gr/ground.h"
